api/requirements.txt
Flask
requests
openai
pymongo
certifi
python-dotenv